requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
moviepy>=2.0.0
openai>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
